# app/cache/test/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Trailburning Journey API",
        "contact": {
            "name": "Matt Allbuery",
            "url": "http://www.trailburning.com/",
            "email": "matt@trailburning.com"
        },
        "version": "2.0"
    },
    "basePath": "/v2",
    "consumes": [
        "application/json",
        "multipart/form-data",
        "application/x-www-form-urlencoded"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/journeys/{id}": {
            "get": {
                "tags": [
                    "Journeys"
                ],
                "summary": "Find a journey by ID",
                "description": "Returns a single journey.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the journey to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Journey"
                        }
                    },
                    "404": {
                        "description": "journey not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Journeys"
                ],
                "summary": "Update a journey",
                "description": "Updates a journey.",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the journey to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "The name of the journey",
                        "type": "string"
                    },
                    {
                        "name": "about",
                        "in": "formData",
                        "description": "About the journey",
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "formData",
                        "description": "The ID of the user the journey belongs to",
                        "type": "string"
                    },
                    {
                        "name": "position",
                        "in": "formData",
                        "description": "The sort position",
                        "type": "integer"
                    },
                    {
                        "name": "publish",
                        "in": "formData",
                        "description": "Publish this journey",
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid data."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Journeys"
                ],
                "summary": "Delete a journey",
                "description": "Deletes the journey.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Journey",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Journey not found"
                    }
                }
            }
        },
        "/journeys/user/{id}": {
            "get": {
                "tags": [
                    "Journeys"
                ],
                "summary": "Find journeys by user",
                "description": "Returns all journeys by a user.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Journey"
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/journeys/{id}/import/gpx": {
            "post": {
                "tags": [
                    "Journeys"
                ],
                "summary": "Import a GPX file",
                "description": "Imports a GPX file and adds the routes found in the GPX to a journey. Replaces all previous routes.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the journey",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "The GPX File to import",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Journey not found"
                    },
                    "400": {
                        "description": "Invalid or empty GPX file."
                    }
                }
            }
        },
        "/journeys/{id}/route_points": {
            "delete": {
                "tags": [
                    "Journeys"
                ],
                "summary": "Delete all route points",
                "description": "Deletes all route points of a journey.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the journey",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Journey"
                        }
                    },
                    "404": {
                        "description": "Journey not found"
                    }
                }
            }
        },
        "/journeys": {
            "post": {
                "tags": [
                    "Journeys"
                ],
                "summary": "Add a journey",
                "description": "Adds a journey.",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "The name of the journey",
                        "type": "string"
                    },
                    {
                        "name": "about",
                        "in": "formData",
                        "description": "About the journey",
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "formData",
                        "description": "The ID of the user the journey belongs to",
                        "type": "string"
                    },
                    {
                        "name": "position",
                        "in": "formData",
                        "description": "The sort position, is handled automatically if not specified",
                        "type": "integer"
                    },
                    {
                        "name": "publish",
                        "in": "formData",
                        "description": "Publish this journey, default value is 'false'",
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation. The Location header contains a link to the new journey.",
                        "headers": {
                            "location": {
                                "description": "Link to the new event.",
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Journey": {
            "required": [
                "id",
                "name",
                "about"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "about": {
                    "type": "string"
                },
                "route_points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RoutePoint"
                    }
                }
            },
            "xml": {
                "name": "Journey"
            }
        },
...
        "RoutePoint": {
            "required": [
                "coords"
            ],
            "properties": {
                "coords": {
                    "description": "point"
                }
            },
            "xml": {
                "name": "RoutePoint"
            }
        }
    }
}